profile subhost 功能完善

+++ Swing/data/src/main/java/com/kidsdynamic/data/net/host/model/SubHostRemovedKidRequest.java
@@ -11,14 +11,14 @@ public class SubHostRemovedKidRequest {
      * kidId : 1
      */
 
-    private int subHostId;
+    private long subHostId;
     private long kidId;
 
-    public int getSubHostId() {
+    public long getSubHostId() {
         return subHostId;
     }
 
-    public void setSubHostId(int subHostId) {
+    public void setSubHostId(long subHostId) {
         this.subHostId = subHostId;
     }
